New architecture and implementation of rules.
ruleast.py: New design of rule AST.

Now the RuleSet is a NameScope and the predicates are names presented by NamedVar.
Predicates work similarly to the other variables except that when adding a name, find the name not only in local scope but from parent scopes as well. So the local variables are only defined in RuleSet.
Conventional properties are defined for each RuleSet so the base predicate, derived predicate and the bounded/unbounded of them can be easily retrieved.

ruleparser.py: the new rule parser.

New architecture of rules and other extensions:
Now the rule parser takes the DistAlgo AST and produces a transformed version of DistAlgo AST that can be fed directly into the Python Generator. In this way, any extensions can be linked easily into DistAlgo.

The new rule parser works in two passings.
The first pass transforms the rules function into rule AST and calls XSB translator to generate the XSB rule files. Each bounded_base predicate is marked with all the rules it triggers and each bounded_derived predicate is marked with the rule that defines it.
The second pass generates implicit infer function calls and completes the arguments of user defined infer calls.

New logic of automatic maintenance of variables: infer at use when needed.
For each derived predicate, define a flag variable indicating if it needs to be updated before use. The flag will be set to True when any related base predicates are assigned (in AssignmentCtx) or updated (in UpdateCtx).
When a derived predicate will be used (in ReadCtx), if its flag of update is True, prepare to call infer on its rule. Before the call, recursively check each of the bounded base predicates of the rule whether its update flag is True and infer them first if necessary. After the infer of a derived predicate, set its update flag to False.

infer.py: simplified infer function

Now the default values of bindings and queries are generated at compile time, so no need to pass in the environment and find the default values in the infer function.

TODO: update the implementation of constraints to the new architecture, so that the exparser and expygen can be removed.