change the syntax of defining rules from: 	def rules(name= ): to: 	def rules_name():
in the first pass of the ruleparser, the rules will be transformed to functions with the same name that calls infer with proper arguments.
in the second pass of the ruleparser, the call of infer will be transformed to directly calling the rules function. and the automatic maintainence calls the rules functions as well.