New architecture, specification and implementation of constraints
1. New architecture
Now the constraint parser takes a DistAlgo AST and a transformed version of DistAlgo AST that can be fed directly into the Python Generator. In this way, any extensions can be linked easily into DistAlgo.

2. New specification
Now the constraint is specified by a function of the form csp_<name>(*parameters),
And the domain specification is similar to the python annotation that use square brackets to define sub-domains

3. New implementation
Now each constraint will be implemented as a function with the same name as it is defined. It will call a builtin _query function to invoke the solvers and process the output.
The output of the constraints will be the corresponding values of the tuple or singleton that is defined at the first place in the return statement.
When using the constraint, the user can directly call the function with defined parameters