Fix infer.
Maybe due to the python version updates, currently, DistAlgo will translate 1 and 0 to True and False, I haven't find the reason yet.
And True and False can also ridiculously pass the test of `isinstance(var, int)`,
So currently when generating the fact, just translate True and False back to 1 and 0.